+++ test/spec/content.adapter.ut.js
@@ -1,16 +1,16 @@
 (function() {
     'use strict';
 
-    define(['app'], function(appModule) {
-        /* global angular:true */
+    define(['app', 'angular'], function(appModule, angular) {
         var copy = angular.copy,
             extend = angular.extend;
 
         describe('ContentAdapter', function() {
             var ContentAdapter,
                 adapter,
-                cinema6,
-                $q;
+                $q,
+                $rootScope,
+                cinema6;
 
             var $httpBackend;
 
@@ -18,6 +18,10 @@
                 module(appModule.name);
 
                 inject(function($injector) {
+                    $rootScope = $injector.get('$rootScope');
+                    $q = $injector.get('$q');
+                    cinema6 = $injector.get('cinema6');
+
                     ContentAdapter = $injector.get('ContentAdapter');
                     ContentAdapter.config = {
                         apiBase: '/api'
@@ -27,9 +31,6 @@
                         config: ContentAdapter.config
                     });
 
-                    cinema6 = $injector.get('cinema6');
-                    $q = $injector.get('$q');
-
                     $httpBackend = $injector.get('$httpBackend');
                 });
             });
@@ -38,13 +39,73 @@
                 expect(adapter).toEqual(jasmine.any(Object));
             });
 
+            describe('decorateWithUser(experience)', function() {
+                var experience, user,
+                    success, failure;
+
+                beforeEach(function() {
+                    experience = {
+                        user: 'u-8da73bf276bb97'
+                    };
+                    user = {
+                        id: 'u-8da73bf276bb97'
+                    };
+
+                    success = jasmine.createSpy('success()'); failure = jasmine.createSpy('failure()');
+
+                    spyOn(cinema6.db, 'find').and.returnValue($q.when(user));
+
+                    $rootScope.$apply(function() {
+                        adapter.decorateWithUser(experience).then(success, failure);
+                    });
+                });
+
+                it('should fetch the user', function() {
+                    expect(cinema6.db.find).toHaveBeenCalledWith('user', 'u-8da73bf276bb97');
+                });
+
+                it('should put the user on the experience', function() {
+                    expect(experience.user).toBe(user);
+                });
+
+                it('should be fulfilled with the experience', function() {
+                    expect(success).toHaveBeenCalledWith(experience);
+                });
+
+                describe('if the service fails to get the user', function() {
+                    beforeEach(function() {
+                        [success, failure].forEach(function(spy) {
+                            spy.calls.reset();
+                        });
+
+                        cinema6.db.find.and.returnValue($q.reject('NOT AUTHORIZED'));
+
+                        $rootScope.$apply(function() {
+                            adapter.decorateWithUser(experience).then(success, failure);
+                        });
+                    });
+
+                    it('should still succeed', function() {
+                        expect(success).toHaveBeenCalledWith(experience);
+                    });
+
+                    it('should set the user to null', function() {
+                        expect(experience.user).toBeNull();
+                    });
+                });
+            });
+
             describe('findAll(type)', function() {
                 var experiences,
-                    users,
                     success;
 
                 beforeEach(function() {
                     success = jasmine.createSpy('success');
+
+                    spyOn(adapter, 'decorateWithUser')
+                        .and.callFake(function(experience) {
+                            return $q.when(experience);
+                        });
 
                     /* jshint quotmark:false */
                     experiences = [
@@ -79,34 +140,11 @@
                             user: "not-e2e-user",
                             org: "not-e2e-org",
                         }
-                    ],
-                    users = [
-                        {
-                            id: "e2e-user",
-                            org: "e2e-org",
-                            status: "active"
-                        },
-                        {
-                            id: "not-e2e-user",
-                            org: "not-e2e-org",
-                            status: "active"
-                        }
                     ];
                     /* jshint quotmark:single */
 
                     $httpBackend.expectGET('/api/content/experiences')
                         .respond(200, experiences);
-
-                    spyOn(cinema6.db, 'find').and.callFake(function(type, user) {
-                        switch (user) {
-                        case 'e2e-user':
-                            return $q.when(users[0]);
-                            break;
-                        case 'not-e2e-user':
-                            return $q.when(users[1]);
-                            break;
-                        }
-                    });
 
                     adapter.findAll('experience').then(success);
 
@@ -114,18 +152,13 @@
                 });
 
                 it('should resolve to all the experiences', function() {
-                    var decoratedExperiences = copy(experiences);
-                    decoratedExperiences.forEach(function(exp) {
-                        switch (exp.user) {
-                        case 'e2e-user':
-                            exp.user = users[0];
-                            break;
-                        case 'not-e2e-user':
-                            exp.user = users[1];
-                            break;
-                        }
+                    expect(success).toHaveBeenCalledWith(experiences);
+                });
+
+                it('should decorate all the experiences with their user', function() {
+                    experiences.forEach(function(experience, index, array) {
+                        expect(adapter.decorateWithUser).toHaveBeenCalledWith(experience, index, array);
                     });
-                    expect(success).toHaveBeenCalledWith(decoratedExperiences);
                 });
             });
 
@@ -146,6 +179,8 @@
                     };
                     /* jshint quotmark:single */
 
+                    spyOn(adapter, 'decorateWithUser').and.returnValue($q.when(experience));
+
                     $httpBackend.expectGET('/api/content/experience/e2e-getid1')
                         .respond(200, experience);
 
@@ -157,6 +192,10 @@
                 it('should return the experience in an array', function() {
                     expect(success).toHaveBeenCalledWith([experience]);
                 });
+
+                it('should decorate the experience with its user', function() {
+                    expect(adapter.decorateWithUser).toHaveBeenCalledWith(experience);
+                });
             });
 
             describe('findQuery(type, query)', function() {
@@ -164,6 +203,11 @@
                     experiences;
 
                 beforeEach(function() {
+                    spyOn(adapter, 'decorateWithUser')
+                        .and.callFake(function(experience) {
+                            return $q.when(experience);
+                        });
+
                     /* jshint quotmark:false */
                     experiences = [
                         {
@@ -195,20 +239,40 @@
                 });
 
                 describe('when there are results', function() {
+                    var meta;
+
                     beforeEach(function() {
+                        meta = {};
+
                         $httpBackend.expectGET('/api/content/experiences?sort=id,1&user=e2e-user')
-                            .respond(200, experiences);
+                            .respond(200, experiences, {
+                                'Content-Range': 'items 10-12/72'
+                            });
 
                         adapter.findQuery('experience', {
                             user: 'e2e-user',
                             sort: 'id,1'
-                        }).then(success);
+                        }, meta).then(success);
 
                         $httpBackend.flush();
                     });
 
+                    it('should decorate the meta object with pagination info', function() {
+                        expect(meta.items).toEqual({
+                            start: 10,
+                            end: 12,
+                            total: 72
+                        });
+                    });
+
                     it('should resolve to the experiences', function() {
                         expect(success).toHaveBeenCalledWith(experiences);
+                    });
+
+                    it('should decorate all the experiences with their users', function() {
+                        experiences.forEach(function(experience, index, array) {
+                            expect(adapter.decorateWithUser).toHaveBeenCalledWith(experience, index, array);
+                        });
                     });
                 });
 
@@ -262,7 +326,6 @@
                         title: "test experience",
                         access: "public",
                         status: "inactive",
-                        user: "e2e-user",
                         org: "784hf785",
                         created: "Blah Blah"
                     };
@@ -270,11 +333,12 @@
 
                     response = extend(copy(experience), { id: 'e-8bf47900eb6fd6' });
 
+                    spyOn(adapter, 'decorateWithUser').and.returnValue($q.when(response));
+
                     $httpBackend.expectPOST('/api/content/experience', {
                         title: 'test experience',
                         access: 'public',
-                        status: 'inactive',
-                        user: 'e2e-user'
+                        status: 'inactive'
                     }).respond(201, response);
 
                     adapter.create('experience', copy(experience)).then(success);
@@ -284,6 +348,10 @@
 
                 it('should respond with the response in an array', function() {
                     expect(success).toHaveBeenCalledWith([response]);
+                });
+
+                it('should decorate the experience with its user', function() {
+                    expect(adapter.decorateWithUser).toHaveBeenCalledWith(response);
                 });
             });
 
@@ -328,13 +396,17 @@
                         created: "fkdslf",
                         lastUpdated: "fkdsjfkd",
                         org: "483fh38",
-                        user: "e2e-user"
+                        user: {
+                            id: "e2e-user"
+                        }
                     };
                     /* jshint quotmark:single */
 
                     response = extend(copy(model), {
                         lastUpdated: 'YASSS'
                     });
+
+                    spyOn(adapter, 'decorateWithUser').and.returnValue($q.when(response));
 
                     $httpBackend.expectPUT('/api/content/experience/e2e-put1', {
                         title: 'origTitle',
@@ -351,6 +423,10 @@
 
                 it('should resolve to the response in an array', function() {
                     expect(success).toHaveBeenCalledWith([response]);
+                });
+
+                it('should decorate the experience with its user', function() {
+                    expect(adapter.decorateWithUser).toHaveBeenCalledWith(response);
                 });
             });
         });